requirements-dev.txt
-r requirements.txt

black
coveralls
django-stubs
mypy
pytest
pytest-cov
pytest-django
pytest-env
pytest-mypy
